requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Session==0.5.0
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
langchain==0.1.0
langchain-community==0.0.10
Werkzeug==2.3.7
gunicorn==21.2.0